Separated, noted profane words
Jesus Christ, this was nonsense. So, highlighting words within a text input is literally impossible in HTML, fun fact. Anyways, the filters I made still don't work quite how I want, since for SOME reason they're dropping the ValidationError and just fixing it to the stringified version sans the code and params. And also have an extra for loop in there for NO reason, which I imagine is related somehow but I honestly don't understand why it's needed but it is. Anyways. I hated this. This took like 5 hours.