Finished\!
Wow that's a lot of code. So. Um. I'll post in Slack and put into Trello the next things that need doing, but I guess the main takeaway for other programmers is to probably not use the SellingForm unless it's associated with an actual user request, because you'd need to create a mock request object with very specific properties to pass in the images associated with the item (it was the best I came up with, feel *free* to take a stab at reengineering if you really want).