Carousal Control Changes
Delete some carousal css on base.css as it was interfering and adding to the css is not doing anything to the carousal control. Added a small black backdrop for the indicator to help with visibility. The next background for buttons are transparents. The next and previous button should only move picture in the carousal is shouldnt do weird stuff.